+++ jars/src/desing/issues
@@ -1,4 +1,4 @@
-MarketRetrofit IndustryChanger:
+MarketRetrofit IndustryChanger (outdated. all issues have been solved?):
     broken instances:
 
         MarketRetrofit_MilitaryBaseInstance:
